Add support for per format packages
