+++ poetry/masonry/utils/include.py
@@ -1,4 +1,5 @@
 from typing import List
+from typing import Optional
 
 from poetry.utils._compat import Path
 
@@ -17,9 +18,12 @@ class Include(object):
         - a directory
     """
 
-    def __init__(self, base, include):  # type: (Path, str) -> None
+    def __init__(
+        self, base, include, formats=None
+    ):  # type: (Path, str, Optional[List[str]]) -> None
         self._base = base
         self._include = str(include)
+        self._formats = formats
 
         self._elements = sorted(list(self._base.glob(str(self._include))))
 
@@ -30,6 +34,10 @@ class Include(object):
     @property
     def elements(self):  # type: () -> List[Path]
         return self._elements
+
+    @property
+    def formats(self):  # type: () -> Optional[List[str]]
+        return self._formats
 
     def is_empty(self):  # type: () -> bool
         return len(self._elements) == 0
