+++ tests/masonry/builders/fixtures/with-include/pyproject.toml
@@ -25,6 +25,8 @@ packages = [
     { include = "extra_dir/**/*.py" },
     { include = "my_module.py" },
     { include = "package_with_include" },
+    { include = "tests", format = "sdist" },
+    { include = "for_wheel_only", format = ["wheel"] },
 ]
 
 include = [
